+++ nn/src/neural_network/feedforward.c
@@ -12,7 +12,7 @@
 #include "neural_network.h"
 #include "utils.h"
 
-NeuralNetwork* create_network(int num_layers) {
+NeuralNetwork* create_network(size_t num_layers) {
   NeuralNetwork* nn = (NeuralNetwork*)malloc(sizeof(NeuralNetwork));
   if (nn == NULL) {
     LOG_ERROR("Memory allocation failed for Neural Network struct.");
@@ -43,7 +43,7 @@ void free_network(NeuralNetwork* nn) {
   }
 
   if (nn->layers != NULL) {
-    for (int i = 0; i < nn->num_layers; i++) {
+    for (size_t i = 0; i < nn->num_layers; i++) {
       if (nn->layers[i] != NULL) {
         if (nn->layers[i]->weights != NULL) {
           free_matrix(nn->layers[i]->weights);
@@ -73,14 +73,14 @@ Matrix* feedforward(NeuralNetwork* nn, const Matrix* input) {
 
   put_matrix(nn->cache, "input", current_output);
 
-  for (int i = 0; i < nn->num_layers; i++) {
+  for (size_t i = 0; i < nn->num_layers; i++) {
     Matrix* z_linear = dot_matrix(current_output, nn->layers[i]->weights);
     // Bias add returns a new matrix; capture it to avoid dropping the result.
     Matrix* z = add_matrix(z_linear, nn->layers[i]->bias);
 
     // Cache the intermediate pre-activation value (z).
     char z_key[32];
-    sprintf(z_key, "z_%d", i);
+    sprintf(z_key, "z_%zu", i);
     put_matrix(nn->cache, z_key, z);
 
     Matrix* a = NULL;
@@ -92,7 +92,7 @@ Matrix* feedforward(NeuralNetwork* nn, const Matrix* input) {
     }
 
     char a_key[32];
-    sprintf(a_key, "a_%d", i);
+    sprintf(a_key, "a_%zu", i);
     put_matrix(nn->cache, a_key, a);
 
     free_matrix(z_linear);
